Handle going downstairs and landing somewhere other than the stairs
The previous code caused the tile we ended up on to be repeatedly
reblessed into upstairs and downstairs alternately, because despite
the bad_staircase message it was assuming that we'd successfully gone
upstairs/downstairs and was trying to link the dungeon graph. This
steals a trick from Planar for checking the success of stair-climbing;
we see to see if we changed level before trying to connect the graph.
This also removes the dungeon graph inconsistency that this mistake
made when the AI attempted to go upstairs immediately after going
downstairs.