Fix a potential IPHD vs. NetHack race condition
The issue here is to do with whether NetHack ends before or after
the flush is called. The existing code assumed the flush happened
first, and it commonly does; but in the case where the NetHack
process was running unusually quickly and managed to end before
the flush took place, it would go into an infinite loop. The flush
is necessary because otherwise the NetHack process might be aborted
before it was done saving all its data, often leading to a corrupted
save file or a hanging NetHack process hogging CPU. This solution
will attempt to flush the IPHD buffers, but gives up after 2 seconds
(on the assumption that NetHack had managed to finish altogether).
The time limit may need to be increased if NetHack is exceptionally
slow at saving, but I don't think it is.