Handle NetHack having lockfile problems when it tries to save
Occasionally, a NetHack installation can get into a mess where stale
lock files for, e.g. the high score table, can end up in its playfield
directory. With the code before this patch, this got worse and worse
as time went on; as NetHack waited for its lockfile to become
available, TAEB just unilaterally killed the process, and NetHack uses
fully-buffered IO to write its various files (such as save files,
bones and dumps) out. The result is that it misses about half of the
final packet of data being written out, leading to a corrupted
/var/games/nethack directory. This patch makes local NetHack
interfaces wait until NetHack's done with its waiting for lockfiles,
finally quitting it if it takes more than 75 seconds total (the worst-
case for a 'correctly working' version of NetHack is 60 seconds). The
user is informed of what's happening meanwhile (via deinitializing the
interface and using print()) so that they don't think that TAEB is
hanging for the minute or so that the process takes.