Remove redundant computation in T::W::T::update
$hadfriendly is a relic from when the item layer and monster layer
were the same layer internally; nowadays, it's unused, therefore
there's no point in computing it in the first place. (It's rather
useful that that computation was well over 99% of T::W::T::update,
and about 30% of T::W::C::update; in theory this could lead to a
gain of about 10% in overall time, but in practice it's less likely
as most AIs would want to look up monster spoilers themselves, and
in such cases update could use a memoized value. However, it's still
an important change, especially as if any AI ever introduces monster
tracking, it'll benefit from the time savings here.)