Replace use constant with its expansion in Ascend and Descend
For some reason, when running under NYTProf the 'use constant' lines
in Ascend and Descend never run, causing an error whenever an AI
tries to go upstairs or downstairs. Replacing the abbreviation with
what it stands for (which is optimised the same way by Perl) allows
such code to function correctly under NYTProf.